moving _layout and its test into _ui_tester_register
