+++ traitsui/testing/tester/_ui_tester_register/tests/test_layout.py
@@ -10,7 +10,7 @@
 #
 import unittest
 
-from traitsui.testing.tester.editors._layout import column_major_to_row_major
+from traitsui.testing.tester._ui_tester_register._layout import column_major_to_row_major
 
 
 class TestLayout(unittest.TestCase):
