+++ traitsui/testing/tester/_ui_tester_register/wx/_traitsui/enum_editor.py
@@ -20,7 +20,7 @@ from traitsui.testing.tester._ui_tester_register._common_ui_targets import (
     _BaseSourceWithLocation
 )
 from traitsui.testing.tester._ui_tester_register.wx import _interaction_helpers
-from traitsui.testing.tester.editors._layout import column_major_to_row_major
+from traitsui.testing.tester._ui_tester_register._layout import column_major_to_row_major
 
 
 class _IndexedListEditor(_BaseSourceWithLocation):
